feat: support stop button
