+++ libs/conversation-input/src/components/Input/Input.tsx
@@ -30,9 +30,9 @@ import { resolveIconUrl } from '../../utils/resolveIconUrl.js';
 import { AttachmentTray } from '../AttachmentTray/AttachmentTray.js';
 import { BottomSheet } from '../BottomSheet/BottomSheet.js';
 import { ModelSelectorBottomSheet } from '../ModelSelectorBottomSheet/ModelSelectorBottomSheet.js';
+import { SendButton } from './Buttons/SendButton.js';
+import { StopButton } from './Buttons/StopButton.js';
 import styles from './Input.module.scss';
-import { SendButton } from './SendButton.js';
-import { StopButton } from './StopButton.js';
 
 export const Input: FC<InputProps> = ({
   message: messageProp = '',
@@ -72,6 +72,7 @@ export const Input: FC<InputProps> = ({
     '--ci-placeholder': colors?.placeholder,
     '--ci-send-bg': colors?.sendBackground,
     '--ci-send-text': colors?.sendText,
+    '--ci-stop-color': colors?.stopColor,
     '--ci-font-family': typography?.fontFamily,
     '--ci-font-size': typography?.fontSize,
     '--ci-font-weight': typography?.fontWeight,
